changes logo font and renames auth screens and nav componnents
